test: cover Dock unread panel composition
